% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ams_ib_ontology.R
\name{read_ams_ib_grid_shapes}
\alias{read_ams_ib_grid_shapes}
\title{Load AMS-I.B grid connection applicability shape}
\usage{
read_ams_ib_grid_shapes()
}
\value{
A \code{sh_shape_graph} as returned by \code{\link[shapeR:read_shacl]{shapeR::read_shacl()}}.
}
\description{
Load AMS-I.B grid connection applicability shape
}
